+++ tests/objects/generics/test_generic_objects.py
@@ -4,7 +4,6 @@ from typing import Generic, List, Optional, TypeVar, Union
 import pytest
 
 import strawberry
-from strawberry.enum import EnumDefinition
 from strawberry.type import StrawberryList, StrawberryOptional, StrawberryTypeVar
 from strawberry.union import StrawberryUnion
 
@@ -671,37 +670,3 @@ def test_federation():
 
     assert field2_copy.python_name == "node_field"
     assert field2_copy.type is str
-
-
-Enum = EnumDefinition(None, name="Enum", values=[], description=None)
-
-
-@strawberry.type
-class TypeA:
-    name: str
-
-
-@strawberry.type
-class TypeB:
-    age: int
-
-
-@pytest.mark.parametrize(
-    "types,expected_name",
-    [
-        ([StrawberryList(str)], "ExampleListStr"),
-        ([StrawberryList(StrawberryList(str))], "ExampleListListStr"),
-        ([StrawberryList(Enum)], "ExampleListEnum"),
-        ([StrawberryUnion("Union", [TypeA, TypeB])], "ExampleUnion"),
-        ([TypeA], "ExampleTypeA"),
-        ([TypeA, TypeB], "ExampleTypeATypeB"),
-    ],
-)
-def test_name_generation(types, expected_name):
-    @strawberry.type
-    class Example(Generic[T]):
-        node: T
-
-    definition = Example._type_definition
-
-    assert definition.get_name_from_types(types) == expected_name
